Move generic schemas outside resource
This is needed to allow importing generic schemas on other parts
of kinto core without breaking the import order.

Related to #1033
